Refactored utils test directory
