+++ frontend/src/utils/__tests__/Hooks/useAutoResetError.test.js
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
-import useAutoResetError from '../Hooks/useAutoResetError.js';
+import useAutoResetError from '../../Hooks/useAutoResetError.js';
 
 describe('useAutoResetError', () => {
     beforeEach(() => {
